Move sum-array-volumes util into separate file
