+++ workers/loc.api/helpers/utils/index.js
@@ -2,8 +2,10 @@
 
 const orderBy = require('./order-by')
 const pushLargeArr = require('./push-large-arr')
+const sumArrayVolumes = require('./sum-array-volumes')
 
 module.exports = {
+  sumArrayVolumes,
   pushLargeArr,
   orderBy
 }
